+++ src/FrameworkServiceProvider.php
@@ -11,6 +11,7 @@ use BrilliantPortal\Framework\OpenApi\SecuritySchemes\apiKey;
 use GoldSpecDigital\ObjectOrientedOAS\Objects\SecurityRequirement;
 use Illuminate\Console\Scheduling\Schedule;
 use Illuminate\Support\Facades\Gate;
+use Illuminate\Validation\Rules\Password;
 use Laravel\Jetstream\Features;
 
 class FrameworkServiceProvider extends PackageServiceProvider
@@ -47,13 +48,22 @@ class FrameworkServiceProvider extends PackageServiceProvider
     {
 
         /**
-         * Teams.
+         * Admins.
          */
-        if (Features::hasTeamFeatures()) {
-            Gate::define('super-admin', function (User $user) {
-                return $user->is_super_admin;
-            });
-        }
+        Gate::define('super-admin', function (User $user) {
+            return $user->is_super_admin;
+        });
+
+        /**
+         * Passwords.
+         */
+        Password::defaults(function () {
+            $rule = Password::min(8);
+
+            return $this->app->isProduction()
+                        ? $rule->mixedCase()->uncompromised()
+                        : $rule;
+        });
 
         /**
          * API.