+++ src/Http/Middleware/SuperAdmin.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace BrilliantPortal\Framework\Http\Middleware;
+
+use Closure;
+use Illuminate\Http\Request;
+
+class SuperAdmin
+{
+    /**
+     * Handle an incoming request.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  \Closure  $next
+     * @return mixed
+     */
+    public function handle(Request $request, Closure $next)
+    {
+        if (! $request->user()->can('super-admin')) {
+            abort(403);
+        }
+
+        return $next($request);
+    }
+}